MLProject/SensorLive/requirements.txt
pandas
numpy
scikit-learn
pymongo
python-dotenv
fastapi
uvicorn
pydantic
pyyaml
joblib
requests
